fix(windows): match the generated registrant's header and symbol names
Flutter's generated plugin registrant includes <location/location_plugin.h>
and calls LocationPluginRegisterWithRegistrar, but the plugin exposed the
_c_api variant. Consolidates the class and the C entry point into
location_plugin.cpp behind include/location/location_plugin.h with the
expected LocationPluginRegisterWithRegistrar symbol.
